requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
amqp==2.5.2               # via kombu
beautifulsoup4==4.8.2
billiard==3.6.2.0         # via celery
boto3==1.12.5
botocore==1.15.5          # via boto3, s3transfer
celery==4.4.0
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
django-clear-cache==0.3
django-configurations==2.2
django-extensions==2.2.8
django-filter==2.2.0
django-model-utils==4.0.0
django-redis==4.11.0
django-simple-history==2.8.0
django-storages==1.9.1
django==2.2.10
docopt==0.6.2             # via notifications-python-client
docutils==0.15.2          # via botocore
envdir==1.0.1
future==0.18.2            # via notifications-python-client
gunicorn==20.0.4
hiredis==1.0.1
idna==2.9                 # via requests
importlib-metadata==1.5.0  # via kombu
jmespath==0.9.4           # via boto3, botocore
kombu==4.6.7              # via celery
lxml==4.5.0
markdown==3.2.1
monotonic==1.5            # via notifications-python-client
notifications-python-client==5.5.1
psycopg2-binary==2.8.4
pyjwt==1.7.1              # via notifications-python-client
python-dateutil==2.8.1
pytz==2019.3              # via celery, django
raven==6.10.0
redis==3.4.1              # via django-redis
requests==2.23.0
s3transfer==0.3.3         # via boto3
six==1.14.0               # via django-configurations, django-extensions, django-simple-history, python-dateutil
soupsieve==2.0            # via beautifulsoup4
sqlparse==0.3.0           # via django
urllib3==1.25.8           # via botocore, requests
vine==1.3.0               # via amqp, celery
whitenoise==5.0.1
zipp==3.0.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
